# backend/docs/swagger/swaggerSolidApi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Marketplace Api",
    "version": "1.0.0",
    "description": "Simple Marketplace Api",
    "contact": {
      "name": "Yaroslav",
      "email": "kamilka13me@gmal.com"
    }
  },
  "servers": [
    {
      "url": "https://localhost:3001/api/",
      "description": "Main local server Api"
    },
    {
      "url": "https://localhost:3002/api/",
      "description": "Main local prod server Api"
    },
    {
      "url": "https://alicesocial.pp.ua:3001/api/",
      "description": "Main dev server Api"
    },
    {
      "url": "https://alicesocial.pp.ua:3002/api/",
      "description": "Main  server Api"
    }
  ],
...
  "paths": {
...
    "/products": {
      "post": {
        "summary": "Creates a new product",
        "tags": [
          "Products"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "required": true,
                    "description": "Product name"
                  },
                  "description": {
                    "type": "string",
                    "required": true,
                    "description": "Product description"
                  },
                  "price": {
                    "type": "number",
                    "required": true,
                    "format": "double",
                    "description": "Product price"
                  },
                  "discount": {
                    "type": "number",
                    "required": true,
                    "format": "int",
                    "description": "Product price"
                  },
                  "category": {
                    "type": "string",
                    "required": true,
                    "description": "Product category"
                  },
                  "quantity": {
                    "type": "number",
                    "required": true,
                    "format": "int64",
                    "description": "Quantity"
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "description": "Product image"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "product": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "typevj;yf": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid request body"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get a list of All products",
        "description": "Get a list of all products according to filters and sorting.",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Limit number of products returned."
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Offset in products list."
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            },
            "description": "Name of the product to search for."
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Filter by product category or product Id."
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "default": "_id"
            },
            "description": "Field to sort by."
          },
          {
            "in": "query",
            "name": "sortDirection",
            "schema": {
              "type": "integer",
              "enum": [
                1,
                -1
              ],
              "default": 1
            },
            "description": "Sort direction, where 1 is ascending and -1 is descending."
          },
          {
            "in": "query",
            "name": "discount",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "discount filter. if 0 off."
          },
          {
            "in": "query",
            "name": "quantity",
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "quantity filter. if 0 off."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of products.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get a product by ID",
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The product ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The product description by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ObjectId format",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Invalid ObjectId format"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Product not found"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Server error",
                    "error": "error"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Product": {
        "type": "object",
        "required": [
          "name",
          "description",
          "price",
          "category",
          "quantity"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the product"
          },
          "description": {
            "type": "string",
            "description": "The description of the product"
          },
          "price": {
            "type": "number",
            "description": "The price of the product"
          },
          "discount": {
            "type": "number",
            "description": "The discount on the product (if any)",
            "default": 0
          },
          "category": {
            "type": "string",
            "description": "The category of the product"
          },
          "quantity": {
            "type": "number",
            "description": "The quantity of the product in stock"
          },
          "views": {
            "type": "number",
            "description": "The number of views of the product",
            "required": false
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "The images URLs of the product"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the product was created"
          }
        },
        "example": {
          "name": "Example Product Name",
          "description": "This is an example product description.",
          "price": 99.99,
          "discount": 10,
          "category": "Example Category",
          "quantity": 100,
          "views": 150,
          "images": [
            "/image1.jpg",
            "/image2.jpg"
          ],
          "created_at": "2023-01-01T00:00:00Z"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}